bug(INVT-CPC-1437): Fix Inventory Api Calls  (#1071)
**JIRA**: https://champlainsaintlambert.atlassian.net/browse/CPC-1437

**Context:**

This PR focuses on refactoring and stabilizing the inventory module
across the API Gateway and frontend. The goal is to centralize all
inventory-related logic into dedicated controllers, ensure frontend
calls are consistent with backend endpoints, and improve maintainability
by consolidating tests.


Does this PR change the .vscode folder in petclinic-frontend?:
No

Reviewers need to check for any changes to the
.vscode folder and add a comment about it to their review comments.

**Changes**
What are the various changes and what other modules do those changes
affect.
This can be bullet point or sentence format.

**Refactored inventory endpoints**

- Moved all inventory-related endpoints into a new controller in a new
V1 folder so they can be used by everyone.

- Added the breaking v2 endpoints into the same controller to ensure
backward compatibility and avoid breaking changes.

**Frontend fixes**

- Removed hardcoded values in the frontend inventory implementation.

- Updated inventory API calls to use the shared axiosInstance for
consistent configuration and request handling.

**Test refactoring**

- Consolidated all inventory-related tests into the
InventoryControllerV2Test class.
-
- Updated the test structure to align with the new controllers and API
endpoints.


Does this use the v2 API?:
No, as mentioned, I removed all v2 API from v2 inventory controller and
put them in v1 if necessary(avoid issue with breaking changes)


Does this add a new communication between services?:
No

---------
