+++ petclinic-frontend/src/features/inventories/api/AddSupplyToInventory.ts
@@ -5,8 +5,14 @@ export const addSupplyToInventory = async (
   inventoryId: string,
   product: ProductRequestModel
 ): Promise<void> => {
-  await axiosInstance.post<void>(
-    `inventories/${inventoryId}/products`,
-    product
-  );
+  try {
+    await axiosInstance.post<void>(
+      `/inventory/${inventoryId}/products`,
+      product,
+      { useV2: false }
+    );
+  } catch (error) {
+    console.error('Error adding supply to Inventory:', error);
+    throw error;
+  }
 };
