+++ petclinic-frontend/src/features/inventories/api/addInventoryType.ts
@@ -5,10 +5,9 @@ export default async function addInventoryType(
   inventoryType: Omit<InventoryType, 'typeId'>
 ): Promise<void> {
   try {
-    await axiosInstance.post<void>(
-      axiosInstance.defaults.baseURL + 'inventories/types',
-      inventoryType
-    );
+    await axiosInstance.post<void>('/inventory/type', inventoryType, {
+      useV2: false,
+    });
   } catch (error) {
     console.error('Error adding inventory type:', error);
     throw error; // Re-throw the error after logging
