+++ petclinic-frontend/src/features/inventories/api/getAllInventoryTypes.ts
@@ -2,8 +2,16 @@ import axiosInstance from '@/shared/api/axiosInstance.ts';
 import { InventoryType } from '@/features/inventories/models/InventoryType.ts';
 
 export async function getAllInventoryTypes(): Promise<InventoryType[]> {
-  const response = await axiosInstance.get<InventoryType[]>(
-    axiosInstance.defaults.baseURL + 'v2/gateway/inventories/types'
-  );
-  return response.data;
+  try {
+    const response = await axiosInstance.get<InventoryType[]>(
+      '/inventory/types',
+      {
+        useV2: false,
+      }
+    );
+    return response.data;
+  } catch (error) {
+    console.error('Error getting all Inventory Type:', error);
+    throw error;
+  }
 }
